pyproject.toml
[project]
name = "busesramdomforest"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "firebase-functions>=0.1.0",
    "firebase-admin>=6.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "joblib>=1.3.0",
    "python-aemet>=0.4.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "urllib3>=2.6.3",
]
[tool.setuptools.packages.find]
include = ["training*", "functions*", "ingestion*"]
